+++ modules/sdxl_styles.py
@@ -3,8 +3,7 @@ import re
 import json
 import math
 
-from modules.util import get_files_from_folder
-
+from modules.path_utils import get_files_from_folder
 
 # cannot use modules.config - validators causing circular imports
 styles_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '../sdxl_styles/'))
